Updated DM

+++ ExpenseProcessingSystem/ExpenseProcessingSystem/Models/Pending/DMFBTModel_Pending.cs
@@ -14,7 +14,7 @@ namespace ExpenseProcessingSystem.Models.Pending
         public int Pending_FBT_MasterID { get; set; }
         public string Pending_FBT_Name { get; set; }
         public string Pending_FBT_Formula { get; set; }
-        public int Pending_FBT_Tax_Rate { get; set; }
+        public float Pending_FBT_Tax_Rate { get; set; }
         public int Pending_FBT_Creator_ID { get; set; }
         public int Pending_FBT_Approver_ID { get; set; }
         public DateTime Pending_FBT_Filed_Date { get; set; }
